requirements.txt
pandas 
numpy
google-cloud-storage
pyyaml
python-dotenv 
matplotlib 
ipykernel
tensorflow
seaborn
scikit-learn
wordcloud